add modifier group component
