+++ packages/soql-builder-ui/src/modules/querybuilder/where/where.html
@@ -13,55 +13,21 @@
     </div>
     <div class="section_selection-container">
       <header class="header">
+        <div class="header__btn--group">
+          <button class="header__btn">AND</button>
+          <button class="header__btn">OR</button>
+        </div>
         <div class="header__labels">
           <!-- i18n -->
           <p>Field</p>
           <p>Operator</p>
           <p>Criteria</p>
         </div>
-        <div class="header__btn--group">
-          <button class="header__btn">AND</button>
-          <button class="header__btn">OR</button>
-        </div>
       </header>
       <section class="modifier__group">
-        <!-- key is index in the where model -->
-        <select
-          class="modifier__item modifier__fields"
-          name="fields"
-          onchange={handleSelectionEvent}
-          data-el-where-field
-        >
-          <option value="" class="placeholder" defualt>Select fields...</option>
-          <template for:each={whereFields} for:item="field">
-            <option key={field} value={field}>{field}</option>
-          </template>
-        </select>
-        <select
-          class="modifier__item modifier__operator"
-          name="operator"
-          onchange={handleSelectionEvent}
-          data-el-where-operator
-        >
-          <option value="=">=</option>
-          <option value="!=">≠</option>
-          <option value="<">&lt;</option>
-          <option value="<=">≤</option>
-          <option value=">">&gt;</option>
-          <option value=">=">≥</option>
-          <!-- i18n -->
-          <option value="starts">starts with</option>
-          <option value="ends">ends with</option>
-          <option value="contains">contains</option>
-        </select>
-        <input
-          class="modifier__item modifier__criteria"
-          name="criteria"
-          type="text"
-          data-el-where-criteria
-          onkeyup={handleSelectionEvent}
-        />
-        <span class="modifier__item pointer"> x </span>
+        <querybuilder-where-modifier-group
+          all-fields={whereFields}
+        ></querybuilder-where-modifier-group>
       </section>
       <footer>
         <button>Add</button>
