+++ packages/soql-builder-ui/src/modules/querybuilder/where/where.ts
@@ -13,30 +13,4 @@ export default class Where extends LightningElement {
   // _local_list of where claused to render
   // API is different than whats rendered
   // add a phantom where group to render, only until the criteria is filled out
-
-  // TODO: we may want to debounce this with RXJS
-  // send the event target info and throttle only on text input?
-  handleSelectionEvent(e) {
-    e.preventDefault();
-    const fieldEl = this.template.querySelector('[data-el-where-field]');
-    const OperatorEl = this.template.querySelector('[data-el-where-operator]');
-    const criteriaEl = this.template.querySelector('[data-el-where-criteria]');
-
-    if (fieldEl.value && OperatorEl.value) {
-      console.log('change event', {
-        field: fieldEl.value,
-        operator: OperatorEl.value,
-        criteria: criteriaEl.value
-      });
-      const whereSelectionEvent = new CustomEvent('whereselection', {
-        detail: {
-          field: fieldEl.value,
-          operator: OperatorEl.value,
-          criteria: criteriaEl.value
-          //index: update or delete
-        }
-      });
-      this.dispatchEvent(whereSelectionEvent);
-    }
-  }
 }
