+++ packages/soql-builder-ui/src/modules/querybuilder/whereModifierGroup/whereModifierGroup.html
@@ -1,0 +1,41 @@
+<template>
+  <!-- key is index in the where model -->
+  <section class="modifier__group">
+    <select
+      class="modifier__item modifier__fields"
+      name="fields"
+      onchange={handleSelectionEvent}
+      data-el-where-field
+    >
+      <option value="" class="placeholder" defualt>Select fields...</option>
+      <template for:each={allFields} for:item="field">
+        <option key={field} value={field}>{field}</option>
+      </template>
+    </select>
+    <select
+      class="modifier__item modifier__operator"
+      name="operator"
+      onchange={handleSelectionEvent}
+      data-el-where-operator
+    >
+      <option value="=">=</option>
+      <option value="!=">≠</option>
+      <option value="<">&lt;</option>
+      <option value="<=">≤</option>
+      <option value=">">&gt;</option>
+      <option value=">=">≥</option>
+      <!-- i18n -->
+      <option value="starts">starts with</option>
+      <option value="ends">ends with</option>
+      <option value="contains">contains</option>
+    </select>
+    <input
+      class="modifier__item modifier__criteria"
+      name="criteria"
+      type="text"
+      data-el-where-criteria
+      onkeyup={handleSelectionEvent}
+    />
+    <span class="modifier__item pointer"> x </span>
+  </section>
+</template>
